Add a GenericFlashNode and use it
Put the code for flash as a subclass
to GenericNode, so all node types with
a flash can utilize it.

Migrate Z1Node and CC2420Node to extend
this class instead.
